Move test dependencies to dev and split the tests in three files
